refactor: util 함수 디렉토리 정리
