+++ src/components/common/Receipt.tsx
@@ -6,8 +6,8 @@ import PerforatedLineBottom from './modal/quaternaryModal/PerforatedLineBottom';
 import { ReactComponent as HorizontalDashedBorder } from '@assets/border/horizontal-dashed-border.svg';
 import { ReactComponent as BankiDad } from '@assets/illust/banki/banki_dad.svg';
 import { ReactComponent as BankiMom } from '@assets/illust/banki/banki_mom.svg';
-import { renderItemIllust } from '@lib/utils/kid';
-import { getContractEndDate } from '@lib/utils/common';
+import { renderItemIllust } from '@lib/utils/common/renderItemIllust';
+import { getContractEndDate } from '@lib/utils/common/getContractEndDate';
 
 interface ReceiptProps {
   createdAt: string;
