+++ src/components/common/button/RoleButton.tsx
@@ -2,7 +2,8 @@ import { HTMLAttributes, useState } from 'react';
 import styled, { css } from 'styled-components';
 import { ReactComponent as RoleButtonBorder } from '@assets/border/role-button-border.svg';
 import { calcRatio, theme } from '@lib/styles/theme';
-import { renderRoleIllust, renderRoleText } from '@lib/utils/kid';
+import { renderRoleIllust } from '@lib/utils/common/renderRoleIllust';
+import { renderRoleText } from '@lib/utils/common/renderRoleText';
 
 interface RoleButtonProps extends HTMLAttributes<HTMLButtonElement> {
   isKid: boolean;
