+++ src/components/common/modal/quaternaryModal/QuaternaryModal.tsx
@@ -9,8 +9,8 @@ import { ReactComponent as BankiMom } from '@assets/illust/banki/banki_mom.svg';
 import { TItemName } from '@lib/types/kid';
 import PerforatedLineTop from './PerforatedLineTop';
 import PerforatedLineBottom from './PerforatedLineBottom';
-import { renderItemIllust } from '@lib/utils/kid';
-import { getContractEndDate } from '@lib/utils/common';
+import { renderItemIllust } from '@lib/utils/common/renderItemIllust';
+import { getContractEndDate } from '@lib/utils/common/getContractEndDate';
 
 interface QuaternaryModalProps {
   /**
