--- src/lib/types/family.tsx
@@ -1,5 +1,0 @@
-export interface IFamilyState {
-  username: string;
-  isFemale: boolean;
-  isKid: boolean;
-}
