+++ src/pages/Home/kid/CreateKid.tsx
@@ -7,12 +7,12 @@ import MarginTemplate from '@components/layout/MarginTemplate';
 import Step3 from '@components/kid/create/content/Step3';
 import Step4 from '@components/kid/create/content/Step4';
 import { useEffect, useState } from 'react';
-import { IFamilyState } from '@lib/types/family';
 import useAxiosPrivate from '@hooks/auth/useAxiosPrivate';
 import Step5 from '@components/kid/create/content/Step5';
 import { useAppDispatch } from '@store/app/hooks';
 import { dispatchParent } from '@store/slices/challengePayloadSlice';
 import { kidMock } from '@lib/mocks/kid';
+import { IFamilyState } from '@lib/types/kid';
 
 const title = [
   '누구와 계약하나요?',
